rapid-api/requirements.txt
Django==4.0.1
djangorestframework== 3.13.1
pytz==2021.3
sqlparse==0.4.2
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2-binary==2.7.7